refactor: move Database into db package
